Make the table of contents only include treatises
This is an attempt to improve the table of contents for the ebook
output. With the questions, it has too much content. Here, we generate a
prologue for each treatise, listing its questions, then we have the
table of contents point to each of these treatise prologues.

I'm not sure if this will work though, as the question content pages
might get skipped because they are not in the table of contents.